Rename OlMapComponent to PlanningMapComponent.
