+++ src/components/map/PlanningMapComponent.vue
